+++ Gemfile
@@ -1,5 +1,3 @@
-
-
 # Gemfile
 
 source 'https://rubygems.org'
