--- ListPeopleOption.rb
@@ -1,14 +1,0 @@
-
-
-require_relative 'menu_option'
-
-# list people
-class ListPeopleOption < MenuOption
-  def execute
-    @app.all_people
-  end
-
-  def name
-    'List all people'
-  end
-end
