+++ base_decorator.rb
@@ -1,6 +1,3 @@
-
-
-# base_decorator.rb
 require_relative 'nameable'
 
 # Base decorator class
