+++ book.rb
@@ -1,6 +1,4 @@
-
-
-# book.rb
+# book class
 class Book
   attr_accessor :title, :author, :rentals
