+++ list_people_option.rb
@@ -1,0 +1,12 @@
+require_relative 'menu_option'
+
+# list people
+class ListPeopleOption < MenuOption
+  def execute
+    @app.all_people
+  end
+
+  def name
+    'List all people'
+  end
+end
