+++ menu_option.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # menu options
 class MenuOption
   def initialize(app)
