+++ nameable.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 # nameable.rb
 class Nameable
   def correct_name
