+++ trim_decorator.rb
@@ -1,6 +1,3 @@
-# frozen_string_literal: true
-
-# trimmer_decorator.rb
 require_relative 'base_decorator'
 
 # trim decorator implements base decorator correct name method
